adding basic file checking so we don't repeat work on repeated runs

+++ rackerkeys.sh
@@ -25,16 +25,29 @@ fi
 # Update packages / package sources
 $PKGUPD update
 
-# Add the rack user
+# Add and configure rack user access
+echo "Adding the Rackspace Management User..."
 useradd -m -d $RACKHOME $RACKUSER
+echo "Configuring SSH keys for access..."
 mkdir $RACKHOME/.ssh
-wget -O $RACKHOME/.ssh/authorized_keys https://raw.githubusercontent.com/rax-brazil/pub-ssh-keys/master/authorized_keys
+wget -sO $RACKHOME/.ssh/authorized_keys https://raw.githubusercontent.com/rax-brazil/pub-ssh-keys/master/authorized_keys
+echo "Correcting SSH configuration permissions..."
 chmod 600 $RACKHOME/.ssh/authorized_keys
 chmod 500 $RACKHOME/.ssh
 chown -R $RACKUSER:$RACKUSER $RACKHOME/.ssh
-echo "*/25 * * * * wget -O /home/rack/.ssh/authorized_keys https://raw.githubusercontent.com/rax-brazil/pub-ssh-keys/master/authorized_keys" > $RACKHOME/rack.cron
-sudo -u $RACKUSER crontab $RACKHOME/rack.cron
 
-echo "# Rackspace user allowed sudo access" > /etc/sudoers.d/rack-user
-echo "$RACKUSER ALL=(ALL) NOPASSWD:ALL" >> /etc/sudoers.d/rack-user
-echo "" >> /etc/sudoers.d/rack-user
+if [ -f $RACKHOME/rack.cron ]; then
+	echo "Crontab already configured for updates...Skipping"
+else
+	echo "Adding crontab entry for continued updates..."
+	echo "*/25 * * * * wget -O /home/rack/.ssh/authorized_keys https://raw.githubusercontent.com/rax-brazil/pub-ssh-keys/master/authorized_keys" > $RACKHOME/rack.cron
+	crontab -u $RACKUSER $RACKHOME/rack.cron
+
+if [ -f /etc/sudoers.d/rack-user ]; then
+	echo "Sudo already configured for Rackspace Management User...Skipping"
+else
+	echo "Configuring sudo for Rackspace Management User"
+	echo "# Rackspace user allowed sudo access" > /etc/sudoers.d/rack-user
+	echo "$RACKUSER ALL=(ALL) NOPASSWD:ALL" >> /etc/sudoers.d/rack-user
+	echo "" >> /etc/sudoers.d/rack-user
+fi
